Corrected a bug which caused an exception when the user tried to delete a list type

+++ server/trunk/KuwaibaServer-ejb/src/java/businesslogic/BackendBean.java
@@ -599,47 +599,51 @@ public class BackendBean implements BackendBeanRemote {
         System.out.println(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_CALL_REMOVEOBJECT"));
 
         if (em != null){
+            RootObject obj = (RootObject)em.find(className, oid);
+            if (className.equals(InventoryObject.class)){ // If the object is an inventory object, we have to delete the children first
 
-            InventoryObject obj = (InventoryObject)em.find(className, oid);
-            if (obj == null)
-                throw new Exception(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_NOSUCHOBJECT")+className+java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_WHICHID")+oid.toString());
+                if (obj == null)
+                    throw new Exception(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_NOSUCHOBJECT")+className+java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_WHICHID")+oid.toString());
 
-            if(obj.isLocked())
-                throw  new Exception(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_OBJECTLOCKED"));
-                       
-            String sentence = "SELECT x FROM ClassMetadata x WHERE x.name ='"+
-                    className.getSimpleName()+"'";
-            //System.out.println(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_EXECUTINGSQL")+sentence);
-            Query query = em.createQuery(sentence);
-            ClassMetadata myClass = (ClassMetadata)query.getSingleResult();
-            List<Object> toBeRemoved = new ArrayList<Object>();
-            for (ClassMetadata possibleChild : myClass.getPossibleChildren()){
-                sentence = "SELECT x FROM "+possibleChild.getName()+" x WHERE x.parent.id="+obj.getId();
+                if(obj.isLocked())
+                    throw  new Exception(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_OBJECTLOCKED"));
+
+                String sentence = "SELECT x FROM ClassMetadata x WHERE x.name ='"+
+                        className.getSimpleName()+"'";
                 //System.out.println(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_EXECUTINGSQL")+sentence);
-                query = em.createQuery(sentence);
-                for (Object removable : query.getResultList()){
-                    InventoryObject myRemovable = (InventoryObject)removable;
-                    //If any of the children is locked, throw an exception
-                    if (!myRemovable.isLocked())
-                        toBeRemoved.add(myRemovable);
-                    else
-                        throw new OperationNotPermittedException(ResourceBundle.getBundle("internationalization/Bundle").
-                                getString("LBL_OBJECTREMOVAL"),ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_OBJECTLOCKED")+
-                                myRemovable.getId()+" ("+myRemovable.getClass()+")");
-                }
-            }
-            toBeRemoved.add(obj);
-            for (Object removable : toBeRemoved){
-                if (removable instanceof ViewableObject){
-                    List<GenericView> views = ((ViewableObject)removable).getViews();
-                    if (views != null){
-                        for (GenericView view : views)
-                            em.remove(view);
-                        views.clear();
+                Query query = em.createQuery(sentence);
+                ClassMetadata myClass = (ClassMetadata)query.getSingleResult();
+                List<Object> toBeRemoved = new ArrayList<Object>();
+                for (ClassMetadata possibleChild : myClass.getPossibleChildren()){
+                    sentence = "SELECT x FROM "+possibleChild.getName()+" x WHERE x.parent.id="+obj.getId();
+                    //System.out.println(java.util.ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_EXECUTINGSQL")+sentence);
+                    query = em.createQuery(sentence);
+                    for (Object removable : query.getResultList()){
+                        InventoryObject myRemovable = (InventoryObject)removable;
+                        //If any of the children is locked, throw an exception
+                        if (!myRemovable.isLocked())
+                            toBeRemoved.add(myRemovable);
+                        else
+                            throw new OperationNotPermittedException(ResourceBundle.getBundle("internationalization/Bundle").
+                                    getString("LBL_OBJECTREMOVAL"),ResourceBundle.getBundle("internationalization/Bundle").getString("LBL_OBJECTLOCKED")+
+                                    myRemovable.getId()+" ("+myRemovable.getClass()+")");
                     }
-                    em.remove(removable);
-                 }
-            }
+                }
+
+
+                toBeRemoved.add(obj);
+                for (Object removable : toBeRemoved){
+                    if (removable instanceof ViewableObject){
+                        List<GenericView> views = ((ViewableObject)removable).getViews();
+                        if (views != null){
+                            for (GenericView view : views)
+                                em.remove(view);
+                            views.clear();
+                        }
+                        em.remove(removable);
+                     }
+                }
+            }else em.remove(obj);
         }
         else 
             throw new EntityManagerNotAvailableException();
